BUG: There was an issue in the end to end test. Fixed.
There was an issue were the end to end test grep'd for a file of
a similar name but not the actual file that was uploaded. So
repeated tests might find an older file not the one under test.
Updated the unit test to remove the uploaded file under test and
to make sure it the specific file is what is being tested not
another.